// src/openapi/models/app-client-summary-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AppClientSummaryDto
 */
export interface AppClientSummaryDto {
    /**
     * 
     * @type {string}
     * @memberof AppClientSummaryDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof AppClientSummaryDto
     */
    name?: string;
}