// src/openapi/models/document-space-app-client-member-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DocumentSpaceAppClientMemberRequestDto
 */
export interface DocumentSpaceAppClientMemberRequestDto {
    /**
     * 
     * @type {string}
     * @memberof DocumentSpaceAppClientMemberRequestDto
     */
    appClientId: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof DocumentSpaceAppClientMemberRequestDto
     */
    privileges: Array<DocumentSpaceAppClientMemberRequestDtoPrivilegesEnum>;
}

/**
    * @export
    * @enum {string}
    */
export enum DocumentSpaceAppClientMemberRequestDtoPrivilegesEnum {
    Write = 'WRITE',
    Membership = 'MEMBERSHIP'
}